fix: 500 when send /login request twice
In the RefreshToken entity class, the token field is marked as a unique key, which is achieved by the @Column(nullable = false, unique = true) annotation, while the RefreshToken entity also contains a one-to-one relationship with the TodoUser. So it's required to delete the existing refresh token before adding a new one.

+++ todo-app-springboot/src/main/java/top/fishg/todoapp/repository/auth/RefreshTokenRepository.java
@@ -10,7 +10,7 @@ import top.fishg.todoapp.model.auth.RefreshToken;
 import java.util.Optional;
 
 @Repository
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Integer> {
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
     Optional<RefreshToken> findByToken(String token);
 
     @Modifying
